Fix self-import error
